requirements.txt
Flask==3.0.3
requests==2.31.0
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
pybit==5.7.0
pybit-unified==0.0.1 